Move transformer and supporting code to `transformers` sub-package
