+++ src/main/java/pl/put/poznan/transformer/logic/transformers/TextTransformerErector.java
@@ -1,4 +1,4 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.transformer.logic.transformers;
 
 import pl.put.poznan.transformer.logic.transforms.*;
