+++ src/main/java/pl/put/poznan/transformer/logic/transformers/Transformer.java
@@ -1,4 +1,4 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.transformer.logic.transformers;
 
 /**
  * The Transformer interface defines the contract for text transformation operations.
