+++ src/main/java/pl/put/poznan/transformer/logic/transforms/CapitalizeTextTransformerDecorator.java
@@ -1,6 +1,6 @@
 package pl.put.poznan.transformer.logic.transforms;
 
-import pl.put.poznan.transformer.logic.Transformer;
+import pl.put.poznan.transformer.logic.transformers.Transformer;
 
 /**
  * The CapitalizeTextTransformerDecorator class is a decorator that capitalizes the first letter of the text.
