+++ src/main/java/pl/put/poznan/transformer/logic/transforms/ExpandAcronymTextTransformerDecorator.java
@@ -1,6 +1,6 @@
 package pl.put.poznan.transformer.logic.transforms;
 
-import pl.put.poznan.transformer.logic.Transformer;
+import pl.put.poznan.transformer.logic.transformers.Transformer;
 
 import java.util.HashMap;
 import java.util.Map;
